Added logic to start, stop and reset the timer.
